pyproject.toml
[project]
name = "ml-research-trends"
version = "0.1.0"
description = "A small toolkit for collecting, embedding, and visualizing ML research papers from Semantic Scholar."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.12",
    "ipykernel>=7.2.0",
    "ollama>=0.6.1",
    "pandas>=2.0.0",
    "requests>=2.32.0",
    "seaborn>=0.13.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.66.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.2.0",
    "scikit-learn>=1.4.0",
    "umap-learn>=0.5.5",
    "numpy>=1.26.0",
    "plotly>=5.20.0",
    "streamlit>=1.33.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ml_research_trends"]
